+++ spec/spec_helper.rb
@@ -9,6 +9,10 @@ require_relative 'support/factories'
 require_relative 'support/mocks'
 
 RSpec.configure do |config|
+  config.before(:suite) do
+    Tinyagent::Migrations.run
+  end
+
   # Enable flags like --only-failures and --next-failure
   config.example_status_persistence_file_path = '.rspec_status'